### 2018-08-20 #906 @whedon
[PRE REVIEW]: qsort R package: A New Tool for Scoring Q-sort Data
**Submitting author:** @davidnsousa (<a href="http://orcid.org/0000-0001-7277-6447">David Sousa</a>)
**Repository:** <a href="https://github.com/joaordaniel/qsort" target ="_blank">https://github.com/joaordaniel/qsort</a>
**Version:** v0.2.0
