### 2018-08-24 #911 @leeper
@maxheld83 Here's the review issue. Above you will find a checklist of issues we would like you to complete related to the submitted software and article. If there's anything missing or if there are changes to the software you would like to see, please open Github issues on the source repository.

Let me know if you have any questions about the process. Thanks, again, for volunteering to review!
