### 2018-08-30 #911 @maxheld83
re: **conflict of interest**, just to keep this on the record:
As I mentioned previously to @leeper I've been a (minor) contributor to the [qmethod](https://github.com/aiorazabala/qmethod) package, and am the author of the yet-in-dev [pensieve](https://github.com/maxheld83/pensieve/) package.
Both are R packages for analysing Q methodological data, and might thus be conceived as "competitors" to the package under review here.
On closer inspection, the overlap in functionality is pretty small / nonexistent.
