### 2018-08-30 #911 @leeper
Thanks for the disclosure, @maxheld83. In my judgment this does not a conflict of interest as defined in [our policy](https://github.com/openjournals/joss/blob/master/COI.md).
