### 2018-08-30 joaordaniel/qsort#1 @maxheld83
streamline paper typesetting
this is a super nitty gritty detail, but it'd be nice to go over the paper and fix some typesetting issues, especially with regard to the proper dashes.

> Generally, Q-sets consist of a large set of items- usually sentences printed on separate cards- that describe personality and behavioral characteristics.

these should be en-dashes, I think `---`.

as per review in https://github.com/openjournals/joss-reviews/issues/911
