### 2018-08-30 joaordaniel/qsort#2 @maxheld83
set broader context for paper
The paper is quite succinct and to the point, but I'm having a bit of a hard time seeing it in the proper context. I'd imagine this might be harder for other readers/users still who are not familiar with q.

In no particular oder, a couple of suggestions:

> Generally, Q-sets consist of a large set of items- usually sentences printed on separate cards- **that describe personality and behavioral characteristics.** These items are sorted into a different number of numerical categories (e.g., 9) based on **their salience as descriptors of an individual** (e.g., 1- most uncharacteristic to 9- most characteristic).
(emphasis added)
 - [ ] Start from a broader focus, explaining what Q generally is and how it is broadly used. The use cases described here (in bold in the above) aren't always, even mostly representative of Q generally speaking. For example, Brown mostly does *political* viewpoints etc. Perhaps you're more focused on a narrower use of Q in psychology (?) here, and it'd be nice to kind of locate and justify that in the bigger scheme of things.

> An individual Q-sort (i.e., the scores of all items descriptive of an individual) can then be compared (correlated) with criteria Q-sorts (description of a hypothetical person at the extreme high end of a domain) to index a score for a particular domain.
> Also, subsets of items can be averaged to derive a scale score [@Block1961, @Block1969, @Block1980, @John2014, @Vaughn2014]. In this paper we present an R package that computes scores from criteria Q-sorts and item scales.
- [ ] Might be worthwhile to explain this in more detail, and to explicate that these are *special* or additional ways of doing Q. Probably due to my ignorance (and not being from psych) I haven't heard about these particular approaches.

as per review https://github.com/openjournals/joss-reviews/issues/911
