### 2018-08-30 joaordaniel/qsort#3 @maxheld83
store datasets as separate objects
I'm not familiar with the practice of storing several datasets in *one* list, and this makes the documentation a tad bid awkward, because the docs then refer to *all* the datasets.

The docs could still be partially reused via `@describeIn` or some such thing.

I might be wrong about this and maybe package authors do this, I've just never seen this.

Feel free to close.

As per https://github.com/openjournals/joss-reviews/issues/911
