### 2018-08-30 joaordaniel/qsort#4 @maxheld83
why include random dataset?
If I understand this correctly `ex_qsort` has been randomly drawn:

> For these examples, items were given random scores (1 to 9; least to most characteristic respectively) following a rectangular distribution

In that case, I'm not sure why it would be included in the package.
Seems a little bit misleading to me, and might also cause weird results because there's only this *one* realisation of random data.

Maybe it's better to provide a function to *generate* such random data at runtime?

as per https://github.com/openjournals/joss-reviews/issues/911
