### 2018-08-30 joaordaniel/qsort#5 @maxheld83
code inversion as -1, 1 or TRUE FALSE
if I understand correctly, the inversion columns are stored as `0` and `1` numerics, which I find a little confusing.

```
pederson_s_inv: num [1:90] 0 0 0 0 0 0 0 0 1 0 ...
```

Shouldn't it be proper logical values *or* -1,1, so that the multiplication works?

as per https://github.com/openjournals/joss-reviews/issues/911
