### 2018-08-30 joaordaniel/qsort#6 @maxheld83
tidy up datasets
perhaps it might make sense to adhere to the tidy data doctrine and clean up the datasets a bit?

> (2) a column with Q-set initials,

that, for example, just seems redundant, because it's the same for every row.

as per https://github.com/openjournals/joss-reviews/issues/911
