### 2018-08-30 joaordaniel/qsort#7 @maxheld83
don't rely on column names for matching
this

> `item1` Column name of x containing item 1 score

seems a little dangerous and hard to use for me.
For example, the later columns in `x` might still be out of order, and it's just a very uncommon way to specify a data set.

Instead, I would suggest, (also per #6) to greatly streamline the data sources, and to *only* accept *matrices* of sorts as `x`, which would have to be in the right order.
Q sorts are properly stored as matrices, because rows and columns really are interchangeable.
Mixing this up inside a dataframe with other columns seems confusing to me.

An alternative might be the (nascent) matrix columns in tibbles, which allow for a peaceful coexistence of matrices inside dfs.

One might also automatically check whether all colnames in `x` are in `qset`, but I think relying on that kind of thing is just too complicated.

as per https://github.com/openjournals/joss-reviews/issues/911
