### 2018-08-30 joaordaniel/qsort#8 @maxheld83
add a vignette
I think a short vignette in addition to, or as part of the paper would greatly help adoption for new users.
Maybe I am just super dense, but it seems much harder than it should be start using this package as is, and to understand its use.

Ideally with a couple of plots, to give a quick impression of what is being done (the transformation from sorts to partial cors and scales).

as per https://github.com/openjournals/joss-reviews/issues/911
