### 2018-08-30 joaordaniel/qsort#9 @maxheld83
expose more granular, widely useful functions
as I understand it, `qsort_score()` can currently *only* be used with the datasets that ship with the package, and only using the *one* monolithic `qsort_score()` function, requiring inputs (`x`) in a very particular form.

In my experience, this may be less than ideal because:

1. overspecific inputs makes it pretty hard for users to use, because they now have to wrangle their data in a somewhat peculiar form to use the package.
2. monolithic functions are hard to maintain in the long run
3. monolithic functions are hard to extend by the author or others.

I would think that something along the lines of factoring out (if not already the case), and exposing the separate functions (ie. inverting, scoring, ...) should then be individual functions, accepting data in pretty generic way, with very few, simple arguments.

In such a more modular / widely useful design, I'd also think that the `qsets` are simply *examples* and that users could quickly use the functions for their own q-sorts and criteria q sorts.

as per https://github.com/openjournals/joss-reviews/issues/911
