### 2018-08-30 #911 @maxheld83
re: **authorship**

it appears that most of the substantive commits are made by @joaordaniel, with @davidnsousa doing the paper and JOSS stuff.
Both are listed as paper authors, so that makes sense.
