### 2018-08-30 joaordaniel/qsort#10 @maxheld83
add automated (unit) tests and validation against published results
both can make it easier to inspire confidence in the results, as well as make it safer for you to develop the package further.

automated tests also appear to be required/suggested by joss.

as per https://github.com/openjournals/joss-reviews/issues/911
