### 2018-08-30 joaordaniel/qsort#11 @maxheld83
add defensive programming techniques
a little bit of input validation might make this safer to use

as per https://github.com/openjournals/joss-reviews/issues/911
