### 2018-08-30 #911 @maxheld83
re: **automated tests**

as far as I can see, there are currently no unit or otherwise automated tests, no cross-validation against published results or input validation.
