### 2018-08-30 joaordaniel/qsort#12 @maxheld83
add community guidelines
as per https://github.com/openjournals/joss-reviews/issues/911
