### 2018-08-30 #911 @maxheld83
I've raised a bunch of separate issues in the source repo, and am done with my review for now.

Here's a bit of overall feedback cc @leeper @joaordaniel @davidnsousa:

My suggestion would be to do **major revisions** before accepting the package (with an important caveat, see below).

These kinds of specialized packages can do a great service to a community of researchers, by making their every day *very* domain specific tasks a lot easier. It's a tremendous and laudable effort to build this kind of package for a community.

That being said, in my opinion, a package also shouldn't be *too* specific in its focus, monolithic in its design or arcane in the way it stores and handles data, because this can cause frustration with users, degrade the overall R experience, guide users away from better, more standard uses of R (ie. around tidy data), and, in the worst case, make a package vulnerable to bugs and reproducibility issues.

This is my chief concern with the package: it seems *very* specialised (on only 4 datasets!). This is also reflected in a **statement of need** which I suspect many users and newcomers can't quite follow, and the example usage and API documentation, both of which are pretty sparse and idiosyncratic at the same time. (See separate issues for detailed suggestions).

As the package stands now, I'm not 100% sure that users wouldn't be better off doing the (relatively simple) calculations "by hand", using more standard tools.

I would think that the community would benefit much more from a more granular package, with small, lean separate functions.
I'd hope that this would also be a way for the work of the submitters to see a greater impact.

----

Now, here's the **caveat** *with regard to my own review*:

I'm pretty familiar and experienced with Q methodology, but have (because of my own ignorance) never heard of this *particular* use/adaptation/extension to Q.
It is entirely possible that the community around these extensions is *so* focused on these 4 datasets (or something else) that the (otherwise excessive) specialisation of this package might be justified.

I am unable to speak to that. So, if the authors @joaordaniel @davidnsousa feel that I'm grossly misunderstanding the use case in this niche of a niche community (Q), which *is entirely possible* (science is weird like that), it might be a good idea to add a second reviewer @leeper.

Hope this is helpful.
