### 2018-08-30 #911 @leeper
Thanks, @maxheld83, for the very thorough review.

What I would suggest at this point is @joaordaniel and @davidnsousa take a look through the issues you've opened and the main points raised in your review here and see what they think they can respond to and address at this point. Please try to address as much of this as you can and ping me on this thread when you have an update ready to for further review. Then I'll have a look through and see how we should proceed, including consulting again with @maxheld83 and/or another reviewer.
