### 2018-09-11 26ac49b
Fixed issues #3, #6 and #12
