+++ R/qsets.R
@@ -1,114 +1,0 @@
-#'Q-sets criteria scores and derived scales
-#'
-#'A list containing four different datasets: Attachment Q-set (aqs; version
-#'3.0), California Child Q-set (ccq); Maternal Behaviour Q-set (mbqs; version
-#'3.1); and Preschool Q-set (pq). In these datasets, column names ending in "_c"
-#'refer to criteria sorts' scores, while column names ending in "_s" refer to
-#'scales' scores.
-#'
-#'@format
-#' \enumerate{
-#'  \item \strong{qsets$aqs} AQS dataset with 90 rows (items)
-#'  and 9 columns.
-#'  \describe{
-#'    \item{item}{Item number from 1 to 90.}
-#'    \item{qset}{aqs. Attachment Q-set (version 3.0) (Waters, 1995).}
-#'    \item{sec_c}{Security criterion scores
-#'    (retrieved from Waters et al., 1995).}
-#'    \item{dep_c}{Dependency criterion scores
-#'    (retrieved from Waters et al., 1995).}
-#'    \item{pederson_s}{Pederson & Moran's (1995) scales
-#'    (retrieved from Waters et al., 1995). comp = compliance; sb = secure base,
-#'    fd = fussy/difficult; epc = enjoys physical contact; as = affective sharing.}
-#'    \item{pederson_s_inv}{Items to be inverted in Pederson and Moran scales.}
-#'    \item{posada_s}{Posada et al.'s (1995) scales (retrieved from Waters et al.,
-#'    1995). sim = smooth interactions with mother; pcm = physical contact with
-#'    mother; ioa = interactions with other adults; pm = proximity to mother.}
-#'    \item{posada_s_inv}{Items to be inverted in Posada et al.'s scales.}
-#'    \item{description}{Item description (retrieved from Waters, 1995).} }
-#'
-#'  \item \strong{qsets$ccq} CCQ dataset with 100 rows (items) and 6 columns.
-#'  \describe{
-#'    \item{item}{Item number from 1 to 100.}
-#'    \item{qset}{ccq. California Child Q-set (Block & Block, 1969).}
-#'    \item{scomp_c}{Social competence criterion scores
-#'    (retrieved from Waters et al., 1985).}
-#'    \item{sest_c}{Self-esteem criterion scores
-#'    (retrieved from Waters et al., 1985).}
-#'    \item{sest_c}{Self-esteem criterion scores
-#'    (retrieved from Waters et al., 1985).}
-#'    \item{egores_c}{Ego-resiliency criterion scores (Block & Block, 1980;
-#'    retrieved from Brian Vaughn, personal communication).}
-#'    \item{egocont_c}{Ego-control criterion scores (Block & Block, 1980;
-#'    retrieved from Brian Vaughn, personal communication).}
-#'    \item{shields_s}{Emotion regulation scale
-#'    (retrieved from Shields and Cicchetti, 1997)}
-#'    \item{shields_s_inv}{Items to be inverted in emotion regulation scale.}
-#'    \item{description}{Item description (retrieved from Waters et al., 1985).} }
-#'
-#'  \item \strong{qsets$mbqs} MBQS dataset with 90 rows (items) and 4 columns.
-#'  \describe{
-#'    \item{item}{Item number from 1 to 90.}
-#'    \item{qset}{mbqs. Maternal Behaviour Q-set (version 3.1) (Pederson et al., 1999).}
-#'    \item{sens_s}{Sensibility criterion scores
-#'    (retrieved from Pederson et al., 1999).}
-#'    \item{description}{Item description
-#'    (retrieved from Pederson et al., 1999).} }
-#'
-#'  \item \strong{qsets$pq} PQ dataset with 72 rows (items) and 6 columns.
-#'  \describe{
[…]
