### 2018-09-12 9471832
Fixed issue #8
