+++ vignettes/qsort-vignette.md
@@ -1,0 +1,301 @@
+---
+title: "qsort R package: A Tool for Scoring Q-sort Data"
+author:
+- "João R. Daniel"
+- "David N. Sousa"
+date: "2018-09-12"
+abstract: |
+  The use of Q-sets to describe subjective views on a specific research domain, in a way suitable for quantitative data analysis, has a long tradition in psychological research. Generally, Q-sets consist of a large set of items- usually sentences printed on separate cards- that describe personality and behavioural characteristics. These items are sorted into a different number of numerical categories (e.g., 9) based on their salience as descriptors of an individual (e.g., 1- most uncharacteristic to 9- most characteristic). An individual Q-sort (i.e., the scores of all items descriptive of an individual) can then be compared (correlated) with criteria Q-sorts (description of a hypothetical person at the extreme high end of a domain) to index a score for a particular domain. Also, subsets of items can be averaged to derive a scale score. In this paper we present an \proglang{R} package that computes scores from criteria Q-sorts and item scales.
+bibliography: bib-vignette.bib
+output:
+  html_document:
+    keep_md: yes
+vignette: >
+  %\VignetteIndexEntry{qsort R package: A Tool for Scoring Q-sort Data
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+
+
+## Introduction
+
+This paper presents a tutorial for an R package qsort to analyse Q-sort data. R [@r2018] is a free software environment for statistical computing (https://cloud.r-project.org/), and R packages are collections of functions and datasets developed by the R community that extend base R functionalities. Due to its open access philosophy, R packages are an important tool for reproducible research. Bellow we describe how to install this package and how to use it to score Q-sort data. For the moment, the qsort package includes descriptions and scoring procedures for four different Q-sets: Attachment Q-set (version 3.0) [@Waters1995a], California Child Q-set [@Block1969], Maternal Behaviour Q-set (version 3.1) [@Pederson1999], and Preschool Q-set (@Baumrind1968, revised by Wanda Bronson). These Q-sets have been used extensively in psychological research and a quick search on google scholar shows that in the past five years these references have been cited 500 plus times. Nevertheless, the core function and datasets of this package can be easily updated to accommodate additional Q-sets. If readers of this paper would like to provide additional Q-set descriptions and scoring procedures, please feel free to contact the authors of qsort. We will gladly update the package to expand its capabilities.
+
+## A Brief Summary of Q-sort Methodology
+  
+  Q-sets are instruments, consisting of a comprehensive large set of items (e.g., sentences, images, photos), used to describe subjective views on a specific research domain in a way suitable for quantitative data analysis [@Block1961; @John2014; @Vaughn2014]. In psychological research a broad range of different Q-sets abound. Their items are usually sentences, printed on separate cards, describing personality and behavioural characteristics. For instance, the California Child Q-set [@Block1969] is a 100 items Q-set that includes descriptions such as: "Is curious and exploring, eager for new experiences" (item 40), "Appears to have a high intellectual capacity" (item 68), or "Has an active fantasy life" (item 97).
+  
+## Computing Scores from Q-sort Data
+
[…]
