+++ DESCRIPTION
@@ -25,8 +25,10 @@ Imports: cowplot,
          stats
 Suggests: devtools,
           knitr,
-          roxygen2
+          roxygen2,
+    rmarkdown
 License: GPL-3
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.1.0
+VignetteBuilder: knitr

[…]
