### 2018-09-12 8907d19
Fixed qsort_length issue + minor corrections
