### 2018-10-25 5787138
changes required for joss paper
