+++ paper.md
@@ -6,11 +6,11 @@ tags:
   - Q-sort scales
   - R
 authors:
-  - name: João R. Daniel
-    orcid: 0000-0001-6609-2014
-    affiliation: 1
   - name: David N. Sousa
     orcid: 0000-0001-7277-6447
+    affiliation: 1
+  - name: João R. Daniel
+    orcid: 0000-0001-6609-2014
     affiliation: 1
 affiliations:
  - name: William James Center for Research, ISPA - Instituto Universitário
@@ -21,20 +21,20 @@ bibliography: paper.bib
 
 # Summary
 
-The use of Q-sets to describe subjective views on a specific research domain, in a way suitable for quantitative data analysis, has a long tradition in psychological research. Generally, Q-sets consist of a large set of items- usually sentences printed on separate cards- that describe personality and behavioral characteristics. These items are sorted into a different number of numerical categories (e.g., 9) based on their salience as descriptors of an individual (e.g., 1- most uncharacteristic to 9- most characteristic). An individual Q-sort (i.e., the scores of all items descriptive of an individual) can then be compared (correlated) with criteria Q-sorts (description of a hypothetical person at the extreme high end of a domain) to index a score for a particular domain. Also, subsets of items can be averaged to derive a scale score [@Block1961, @Block1969, @Block1980, @John2014, @Vaughn2014]. In this paper we present an R package that computes scores from criteria Q-sorts and item scales.
+The use of Q-sets to describe subjective views on a specific research domain, in a way suitable for quantitative data analysis, has a long tradition in developmental psychology research. In this field, Q-sets generally consist of a large set of items- usually sentences printed on separate cards- that describe personality and behavioral characteristics. These items are sorted into a different number of numerical categories (e.g., 9) based on their salience as descriptors of an individual (e.g., 1- most uncharacteristic to 9- most characteristic). An individual Q-sort (i.e., the scores of all items descriptive of an individual) can then be compared (correlated) with criteria Q-sorts (description of a hypothetical person at the extreme high end of a domain) to index a score for a particular domain. Also, subsets of items can be averaged to derive a scale score [@Block1961, @Block1969, @Block1980, @John2014, @Vaughn2014]. In this paper we present an R package that computes scores from criteria Q-sorts and item scales as traditionally used in developmental psychology research.
 
-The ``qsort`` package includes descriptions and scoring procedures for four different Q-sets: Attachment Q-set (version 3.0) [@Waters1995], California Child Q-set [@Block1969], Maternal Behaviour Q-set (version 3.1) [@Pederson1999], and Preschool Q-set [@Baumrind1968] revised by Wanda Bronson).
+The ``qsort`` package includes descriptions and scoring procedures for four commonly used Q-sets: Attachment Q-set (version 3.0) [@Waters1995], California Child Q-set [@Block1969], Maternal Behaviour Q-set (version 3.1) [@Pederson1999], and Preschool Q-set [@Baumrind1968] revised by Wanda Bronson).
 
-The package contains four objects: 
+The package contains seven objects: 
 
 * `ex_qsort` - example Q-sort datasets for the four different Q-sets: aqs, ccq, mbqs, pq.
-* `qsets` - a list containing the four different datasets mentioned above.
-* `qsort_score` -  a function which uses the criteria scores and/or scales from the `qsets` datasets to compute scores for Q-sort data.
-* `print_cards`  - a function which creates a .pdf file where items' descriptions are printed in separate cards.
+* `qsets_aqs`, `qsets_ccq`, `qsets_mbqs`, and `qsets_pq` - four different datasets containing items descriptions, criteria socres and/or scales for the corresponding Q-set.
+* `qsort_score` -  a function to compute scores for Q-sort data using the criteria scores and/or scales from the **qsets** datasets.
[…]
