+++ cran-comments.md
@@ -1,3 +1,7 @@
+### Update package 2018-10-25 - qsort 0.2.3
+* qsets datasets now appear as different objects
+* qsort_score function changed accordingly
+
 ### Update package 2018-08-12 - qsort 0.2.2
 * added Shields and Cicchetti’s (1997) emotion regulation scale to qsets$ccq
 * added ORCID number of co-author David N. Sousa in the DESCRIPTION file

[…]
