### 2018-10-25 be42dc7
Changes for JOSS
