### 2018-10-25 6f5880e
fixed warnings and notes for cran
