+++ DESCRIPTION
@@ -27,6 +27,7 @@ Suggests: devtools,
           knitr,
           roxygen2,
     rmarkdown
+Depends: R (>= 2.10)
 License: GPL-3
 Encoding: UTF-8
 LazyData: true
