### 2018-10-26 37e14af
fixes for CRAN
