+++ vignettes/bib-vignette.bib
@@ -10,7 +10,7 @@
   author={Jason Becker and Chung Hong Chan and Geoffrey CH Chan and Thomas J. Leeper and Christopher Gandrud and Andrew MacDonald and Ista Zahn and Stanislaus Stadlmann and Ruaridh Williamson},
   year={2018},
   note = {\proglang{R} package version 0.5.10},
-  url={https://cran.r-project.org/web/packages/rio/index.html}
+  url={https://CRAN.R-project.org/package=rio}
 }
 
 @book{Block1961,
@@ -77,7 +77,7 @@
 @software{r2018,
   author = {{R Core Team}},
   title = {R: A language and environment for statistical computing (Version 3.5.0)},
-  url = {https://cloud.r-project.org/},
+  url = {https://CRAN.R-project.org},
   year = {2018}
 }
 
@@ -132,7 +132,7 @@
   author={Wickham, H. and Fran\c{c}ois, R. and Henry, L. and M\"{u}ller, K.},
   year={2018},
   note = {\proglang{R} package version 0.7.5},
-  url={https://cran.r-project.org/web/packages/dplyr/}
+  url={https://CRAN.R-project.org/package=dplyr}
 }
 
 @misc{wickham2017data,
