### 2018-10-26 joaordaniel/qsort#12 @joaordaniel
done as requested
