### 2018-10-26 joaordaniel/qsort#8 @davidnsousa
Fixed this issue. Html `Vignette` added in vignettes.
