### 2018-10-26 joaordaniel/qsort#6 @davidnsousa
Fixed this issue according to tidy data doctrine.
