### 2018-10-26 joaordaniel/qsort#3 @davidnsousa
Fixed this issue. Q-sets datasets are now stored in separate objects: qset_aqs, qset_ccq, qset_mbqs and qset_pq.
