### 2018-10-26 joaordaniel/qsort#9 @joaordaniel
As suggested the qsort_score() has been edited so it can used with other qsets not shipped with the package.

As to the format of input data we have not changed it because we do not feel it to be particularly peculiar. It is simply a dataset where each column represents the score of each qset item. From our experience, this is the most straightforward and commonly used way to enter these type of data into a spreadsheet.

As to the breaking the qsort_score() into a set of smaller functions we feel it to be a little counter-intuitive to the package's objectives- a fast, little-code required procedure of scoring qsort data as commonly used by developmental psychologists.
