### 2018-10-26 joaordaniel/qsort#7 @joaordaniel
All dataset inputs of every function have to follow some set of rules. Here we only require qset items' columns to be in order from 1 to etc... We do not feel this to be such an uncommon specification, when considering the type of data and instrument these data refer to.
