### 2018-10-26 joaordaniel/qsort#5 @joaordaniel
Inversion columns can be specified in different ways and inverting item scores in scales does not equal to multiplying the score by -1. Inverted/reversed item scores correspond to subtracting item scores from max + 1, so that all scores have the same direction.
