### 2018-10-26 joaordaniel/qsort#4 @joaordaniel
We included a random data set just to have an example to show the kind of output the function creates.
