### 2018-10-26 joaordaniel/qsort#2 @joaordaniel
We edited sections of the paper so that it is clearer now that the package analyses qsort data as it is commonly done in developmental psychology.
