### 2018-10-26 joaordaniel/qsort#11 @davidnsousa
[Please see response two issue #7](https://github.com/joaordaniel/qsort/issues/7#issuecomment-433448220)
