### 2018-10-26 c053706
fixed issues #1 and #10
