+++ paper.md
@@ -21,7 +21,7 @@ bibliography: paper.bib
 
 # Summary
 
-The use of Q-sets to describe subjective views on a specific research domain, in a way suitable for quantitative data analysis, has a long tradition in developmental psychology research. In this field, Q-sets generally consist of a large set of items- usually sentences printed on separate cards- that describe personality and behavioral characteristics. These items are sorted into a different number of numerical categories (e.g., 9) based on their salience as descriptors of an individual (e.g., 1- most uncharacteristic to 9- most characteristic). An individual Q-sort (i.e., the scores of all items descriptive of an individual) can then be compared (correlated) with criteria Q-sorts (description of a hypothetical person at the extreme high end of a domain) to index a score for a particular domain. Also, subsets of items can be averaged to derive a scale score [@Block1961, @Block1969, @Block1980, @John2014, @Vaughn2014]. In this paper we present an R package that computes scores from criteria Q-sorts and item scales as traditionally used in developmental psychology research.
+The use of Q-sets to describe subjective views on a specific research domain, in a way suitable for quantitative data analysis, has a long tradition in developmental psychology research. In this field, Q-sets generally consist of a large set of items, usually sentences printed on separate cards, that describe personality and behavioral characteristics. These items are sorted into a different number of numerical categories (e.g., 9) based on their salience as descriptors of an individual (e.g., 1- most uncharacteristic to 9- most characteristic). An individual Q-sort (i.e., the scores of all items descriptive of an individual) can then be compared (correlated) with criteria Q-sorts (description of a hypothetical person at the extreme high end of a domain) to index a score for a particular domain. Also, subsets of items can be averaged to derive a scale score [@Block1961, @Block1969, @Block1980, @John2014, @Vaughn2014]. In this paper we present an R package that computes scores from criteria Q-sorts and item scales as traditionally used in developmental psychology research.
 
 The ``qsort`` package includes descriptions and scoring procedures for four commonly used Q-sets: Attachment Q-set (version 3.0) [@Waters1995], California Child Q-set [@Block1969], Maternal Behaviour Q-set (version 3.1) [@Pederson1999], and Preschool Q-set [@Baumrind1968] revised by Wanda Bronson).
 
