### 2018-10-26 #911 @joaordaniel
We would like to thank @maxheld83 for his extensive review of the package. We have made some edits to the package to accommodate reviewer's concerns, but chose to maintain some of the code as it was. We present our motives to do so in the replies to specific issues.

We agree with the reviewer that this is a very specific package. It is mostly directed to researchers coming from developmental psychology- we have stressed this in the new version of the paper. For now the package only includes descriptions and scoring options for 4 qsets. These were chosen because there are the most commonly used in developmental psychology studies.

Although its focus is not very broad the package streamlines the data analysis required for researchers in this field. Calculations are quite simple, but when you have a lot of individual data, coming from different classrooms, observers etc., this process can became cumbersome. The functions we share here have made our life a lot easier, and we think researchers using qsets as we do will find this tool helpful as well.

Hope to hear your feedback on this revised version of the package.
