### 2018-11-20 #911 @arfon
@joaordaniel thanks for following up here. Could you be a little more specific about the edits/updates that you _have_ made based on @maxheld83's feedback?
