### 2018-11-20 #911 @arfon
Also :wave: folks. I'll be talking over as editor for this submission since @leeper has unfortunately had to resign his role with JOSS because of his new job.
