### 2018-11-26 #911 @davidnsousa
Dear Arfon @arfon ,

Thank your for getting back to us on our submission.

Reviewer Maximilian Held (@maxheld83) suggested some helpful changes to the github repository and the qsort package. Namely, adding community guidelines, automated units tests, and a vignnete to make the usage of the package more clear for users. We have revised our package accordingly and integrated proposed changes.

Also, Maximilian Held suggested we should change the way that the datasets, required to run our core function (qsort function), were organized. As suggested these datasets were tidied and stored as separate objects-the former list 'qsets' was broken down to 4 different objects (qset_ccq; qset_aqs; qset_mbqs; qset_pq). qsort functions’ code was edited to accommodate these changes.

Maximilian Held also suggested with edited the paper to better explain readers how this package could help them analyse Qsort data. We tried to give a more precise focus on the goals of this package- analyse Qsort data as is commonly done in Developmental Psychology studies. Consequently, we fell the manuscript is now more clear and better shows its potential merits.

Despite these helpful suggestions by Maximilian Held, we believe that changing our code to accommodate two other suggestions, rose in issues 7 and 11, to be a little counter-intuitive to the package's objectives- a fast, little-code required procedure for scoring Qsort data as commonly collected by developmental psychologists. As such we have not broken the qsort function into a number of “smaller” functions, nor the way the input data is specified. Considering the type of instrument used, and how data is stored by researchers using it we do not feel requested changes to be helpful for the users this package may appeal. Our only requirement for input data is that qset item’s columns/variables (scores of these columns are used to computed a set of new variables) to be in order from 1 to max number of qset items.

We hope the changes we made have made our manuscript and package reach the standards of JOSS.
