### 2018-12-01 79c5ca8
Update paper.md
