### 2018-12-01 #911 @arfon
Hi @davidnsousa, thanks for this detailed summary of your changes. Firstly, one change I need you to make to the paper to fix your references formatting: https://github.com/joaordaniel/qsort/pull/13

I think @maxheld83's suggestions in https://github.com/joaordaniel/qsort/issues/11 and https://github.com/joaordaniel/qsort/issues/7 are good ones in fact, we've recently updated our submission requirements to better fit with feedback such as given in this issue: https://github.com/joaordaniel/qsort/issues/9

> - The software should be feature complete (no half-baked solutions) and designed for maintainable extension (not one-off modifications).

This paper was submitting before these guidelines were updated so I don't think it's reasonable to hold this submission to exactly these requirements.

The primary goal of JOSS is to enable academic credit for authors of (open source) research software, a secondary (but very important) goal is to improve the quality of the software being produced by the research community.

This means that sometimes when reviewers and authors disagree on how a package should be constructed, we defer to the author. On balance, I think this is probably one of those times.

@maxheld83 - thanks for all of your efforts thus far. Could you take a look at the un-checked items on your reviewer checklist and see if you're able to verify some of these now?
