### 2018-12-01 joaordaniel/qsort#10 @maxheld83
I think by some standards, these unit tests might be a little shallow in coverage.
Also, there isn't any validation against published results.
