### 2018-12-01 #911 @maxheld83
@arfon thanks.

I've checked the remaining items off of my reviewer checklist above, though I still have reservations concerning the usability, maintainability and focus of the package, as I mentioned above. Among other things, I still think the package is too monolithic and idiosyncratic in its design, and does not follow (enough) software dev best practices, including defensive programming or reasonably extensive unit tests.

But I might be mistaken about the acceptance guidelines of JOSS, and @arfon's decision seems very reasonable to me.

In any event, I wouldn't want to spring such (contentious) demands on @davidnsousa  and @joaordaniel's generous work without it being a transparent JOSS  acceptance criteria.

I added a separate [issue over at joss](https://github.com/openjournals/joss/issues/469) to discuss the criteria.
