### 2018-12-01 #911 @davidnsousa
@arfon thank you for your fix. Merged!

@arfon @maxheld83 thank you both for your comments.
