### 2018-12-02 #911 @arfon
@davidnsousa - I wonder if there's a way to truncate the output of your examples so they don't overflow the margins? We'll accept now and can tweak the appearance later if you want to.
