### 2018-12-02 #911 @arfon
@maxheld83 - many thanks for your review here ✨

@davidnsousa @joaordaniel - your paper is now accepted into JOSS :zap::rocket::boom:
