### 2018-12-02 #911 @davidnsousa
@arfon @maxheld83 thank you both!

I fixed the margins issue. I attempted to generate a new pdf... I hope it s fixed.
